fix: wrong assignment of position and origin
