+++ src/MotivePosition.py
@@ -1,15 +1,7 @@
 from dataclasses import dataclass
-from enum import Enum
 
-
-class PositionInWork(str, Enum):
-    ORIGINAL = 0
-    INVERTED = 1
-    MIRRORED = 2
-    OUTSIDE = 3
-
-    def __str__(self):
-        return self.name
+from MotiveUnit import Origin
+from PositionInWork import PositionInWork
 
 
 @dataclass
@@ -17,21 +9,20 @@ class MotivePosition:
     position: int
     length: int
 
-    original_work: str
-    original_position: int
+    origin: Origin
     position_in_work: PositionInWork
 
     def __str__(self):
-        return f"{self.position}:{self.length}:{self.original_position}:{self.position_in_work}"
+        return f"{self.position}:{self.length}:{self.origin}:{self.position_in_work}"
 
     def __repr__(self):
-        return f"{self.original_position}:{self.length}"
+        return f"{self.origin}:{self.length}"
 
     def is_same_position(self, other: "MotivePosition") -> bool:
         if self.length != other.length:
             return False
 
-        if self.original_work != other.original_work:
+        if self.origin != other.origin:
             return False
 
         if (
@@ -42,15 +33,17 @@ class MotivePosition:
 
         if self.position_in_work is PositionInWork.MIRRORED:
             if other.position_in_work is PositionInWork.MIRRORED:
-                return self.original_position == other.original_position
+                return self.origin.note_number == other.origin.note_number
             else:
                 return (
-                    self.original_position == other.original_position + other.length - 1
+                    self.origin.note_number
+                    == other.origin.note_number + other.length - 1
                 )
         else:
             if other.position_in_work is PositionInWork.MIRRORED:
                 return (
-                    self.original_position + self.length - 1 == other.original_position
+                    self.origin.note_number + self.length - 1
+                    == other.origin.note_number
                 )
             else:
-                return self.original_position == other.original_position
+                return self.origin.note_number == other.origin.note_number
